# Faster GPU NMS operator