## Description ##
This PR significantly improves the performance of `mx.sym.contrib.box_nms` on GPU.

@zhreshold @Jerryzcn FYI

Test cases:

| input size | topk | old time | new time | speedup |
| ------------ | ---- | --- | --- | --- |
| (1, 507, 5) | 507 | 0.78 ms | 0.14 ms | 5.57 |
| (1, 1875, 5) | 1875 | 2.34 ms | 0.22 ms | 10.63 |
| (1, 7500, 5) | 2400 | 4.07 ms | 0.41 ms | 9.93 |
| (1, 30000, 5) | 2400 | 4.17 ms | 0.41 ms | 10.17 |
| (1, 120000, 5) | 2400 | 5.81 ms | 0.47 ms | 12.36 |
| (1, 159882, 5) | 12001 | 10.48 ms | 1.17 ms | 8.96 |

## Checklist ##
### Essentials ###
Please feel free to remove inapplicable items for your PR.
- [x] Changes are complete (i.e. I finished coding on this PR)
- [ ] All changes have test coverage:
- Unit tests are added for small changes to verify correctness (e.g. adding a new operator)
- [ ] Code is well-documented:  
- For new C++ functions in header files, their functionalities and arguments are documented. 
- [x] To the my best knowledge, examples are either not affected by this change, or have been fixed to be compatible with this change

## Comments ##

 - Currently the new function is called when there is no requirement for backward pass, but it is easily extendable to include that.