@ptrendx 2019-10-31
@zhreshold Do you have any other comments?

For the record - I think there are a few outstanding items to fully mark GPU implementation of NMS as "done", namely
 - handle the output required for backward in the new implementation
 - change the way sorting is done (currently sorting happens `batch size` times, and it turns out that for large topk it is not beneficial, the previous implementation that would always do 2 (larger) sorts was better in that regard.
That said, I will not be able to finish those 2 outstanding items before 1.6 code freeze and I believe performance improvements are enough to merge this PR and address those 2 points as part of the future PR.